+++ app/src/screens/Search/index.js
@@ -1,15 +1,6 @@
 import React, {Component} from 'react';
 import {Content, List, ListItem, Text} from 'native-base';
-import {
-  Dimensions,
-  FlatList,
-  Keyboard,
-  Linking,
-  Platform,
-  SafeAreaView,
-  TouchableOpacity,
-  View
-} from 'react-native';
+import {Dimensions, FlatList, Keyboard, Linking, Platform, SafeAreaView, TouchableOpacity, View} from 'react-native';
 
 import defaultStyles from './styles';
 import {FavoriteListItem, PlaceListItem} from '../../components/ListItems';
@@ -60,7 +51,7 @@ export default class SearchScreen extends Component {
   _renderPlaceItem = (place) => {
     return <PlaceListItem place={place} onPress={() => this._onPress(place)}
                           addFavorite={() => this.props.addFavorite(place)}
-                          remFavorite={() => this.props.remFavorite(place)}/>
+                          remFavorite={() => this.props.remFavorite(place)}/>;
   };
   _renderList = () => {
     let data;
@@ -81,12 +72,12 @@ export default class SearchScreen extends Component {
       return (
         <Content>
           <List>
-            <ListItem itemHeader first style={{paddingBottom: favorites.length > 0 ? 0 : 10}}>
+            <ListItem itemHeader first style={[{paddingBottom: favorites.length > 0 ? 0 : 10}, defaultStyles.listHeader]}>
               <Text style={textStyle1}>Favorites</Text>
             </ListItem>
             <Text style={[textStyle2, {display: favorites.length > 0 ? 'none' : 'flex'}]}> No favorites yet </Text>
             {favorites}
-            <ListItem itemHeader first style={{paddingTop: 50, paddingBottom: lastSearches.length > 0 ? 0 : 10}}>
+            <ListItem itemHeader first style={[{paddingBottom: lastSearches.length > 0 ? 0 : 10}, defaultStyles.listHeader, defaultStyles.listHeaderNotFirst]}>
               <Text style={textStyle1}>Last searches</Text>
             </ListItem>
             <Text style={[textStyle2, {display: lastSearches.length > 0 ? 'none' : 'flex'}]}> No last searches
@@ -112,7 +103,7 @@ export default class SearchScreen extends Component {
   render() {
     return (
       <SafeAreaView style={defaultStyles.safeArea}>
-        <View style={[defaultStyles.container]} pointerEvents="box-none">
+        <View style={[defaultStyles.container]} pointerEvents='box-none'>
           {this._renderSearchBar()}
           {this._renderList()}
         </View>