+++ app/RELEASE_NOTES.md
@@ -4,3 +4,7 @@
 * Uses sensible relative styling rules, dealing with the button width in the MapCard.
 * Fixes an floating point issue where the number of seconds in the MapCard would be shown with way too many decimal places.
 * Adds padding for home indicator to MapCard.
+* Removes small lines under 'favorites' and 'searches'.
+* Adds margin above first header in SearchScreen.
+* Adds margin above search field in SearchScreen.
+